@@ -6,10 +6,13 @@ repository: https://gitlab.com/arrow/flutter_reified_lenses
 publish_to: none
 
 environment:
-  sdk: '>=2.6.0 <3.0.0'
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
   reified_lenses:
     path: ../reified_lenses
   flutter:
     sdk: flutter
+
+dev_dependencies:
+  pedantic: ^1.10.0-nullsafety.3
