@@ -6,8 +6,11 @@ repository: https://gitlab.com/arrow/reified_lenses
 publish_to: none
 
 environment:
-  sdk: ^2.8.0
+  sdk: ">=2.12.0-0 <3.0.0"
   # flutter: '>=1.9.0 <1.13.0'
 
 dependencies:
-  meta: ^1.1.8
+  meta: ^1.3.0-nullsafety.6
+
+dev_dependencies:
+  pedantic: ^1.10.0-nullsafety.3
