@@ -6,18 +6,18 @@ repository: https://gitlab.com/arrow/reified_lenses_builder
 publish_to: none
 
 environment:
-  sdk: '>=2.6.0 <3.0.0'
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
-  meta: ^1.1.8
+  meta: ^1.3.0-nullsafety.6
   reified_lenses:
     path:
       ../reified_lenses
   build: '>=0.12.6 <2.0.0'
   source_gen: '>=0.9.0 <1.0.0'
-  analyzer: ^0.39.4
+  analyzer: ^0.41.1
 
 dev_dependencies:
-  build_runner:
-  pedantic: ^1.9.0
+  build_runner: ^1.10.7
+  pedantic: ^1.10.0-nullsafety.3
   effective_dart: ^1.2.1
