@@ -6,12 +6,12 @@ part 'test.g.dart';
 class Data<A> {
   static Lens<Data<A>, Data<A>> lens<A>() => Lens.identity();
 
-  final A a;
-  final Data<A> b;
+  final A? a;
+  final Data<A>? b;
   @mutater
-  final int c;
+  final int? c;
   @getter
-  final String d;
+  final String? d;
   @getter
   String get e => 'this';
 
@@ -19,4 +19,5 @@ class Data<A> {
 }
 
 Data<int> d = Data<int>();
-final a = Data.lens<int>().b.b.a.get(d);
+int? a = Data.lens<int>().b.nonnull.b.nonnull.a.get(d);
+
