@@ -1,6 +1,6 @@
 class Path<E> {
-  final Path<E> _prev;
-  final E _elem;
+  final Path<E>? _prev;
+  final E? _elem;
 
   const Path(this._elem, this._prev)
       : assert(_elem != null),
@@ -15,27 +15,37 @@ class Path<E> {
   Path<E> operator +(Path<E> other) {
     if (other.isSingleton) return Path(other._elem, this);
     if (other.isEmpty) return this;
-    return Path(other._elem, this + other._prev);
+    return Path(other._elem, this + other._prev!);
   }
 
   bool get isEmpty => _elem == null;
-  bool get isSingleton => _prev.isEmpty;
+  bool get isSingleton => _prev?.isEmpty ?? false;
 
   Path<E> _reverse() {
-    if (this.isEmpty || this.isSingleton) return this;
-    var inverted = Path.singleton(this._elem);
-    var rest = this._prev;
-    while (!rest.isEmpty) {
+    if (isEmpty || isSingleton) return this;
+    var inverted = Path.singleton(_elem);
+    var rest = _prev;
+    while (!rest!.isEmpty) {
       inverted = Path(rest._elem, inverted);
       rest = rest._prev;
     }
     return inverted;
   }
+
+  E get first {
+    assert(!isEmpty);
+    return _elem!;
+  }
+
+  Path<E> get rest {
+    assert(!isEmpty);
+    return _prev!;
+  }
 }
 
 class PathMap<K, V> {
-  Set<V> _values = Set.identity();
-  Map<K, PathMap<K, V>> _children = {};
+  final Set<V> _values = Set.identity();
+  final Map<K, PathMap<K, V>> _children = {};
 
   PathMap.empty();
 
@@ -46,29 +56,31 @@ class PathMap<K, V> {
       _values.add(value);
     } else {
       _children
-          .putIfAbsent(key._elem, () => PathMap.empty())
-          ._add(key._prev, value);
+          .putIfAbsent(key.first, () => PathMap.empty())
+          ._add(key.rest, value);
     }
   }
 
   void remove(Path<K> key, V value) => _remove(key._reverse(), value);
 
   void _remove(Path<K> key, V value) {
-    if (key.isEmpty)
+    if (key.isEmpty) {
       _values.remove(value);
-    else {
-      _children[key._elem]?._remove(key._prev, value);
-      if (_children[key._elem]?._children?.isEmpty ?? false)
+    } else {
+      _children[key.first]?._remove(key.rest, value);
+      if (_children[key.first]?._children.isEmpty ?? false) {
         _children.remove(key._prev);
+      }
     }
   }
 
-  Iterable<V> children([Path<K> key]) => _childrenInternal(key);
+  Iterable<V> children([Path<K> key = const Path.empty()]) => _childrenInternal(key);
 
-  Iterable<V> _childrenInternal([Path<K> key]) {
-    if (key != null && !key.isEmpty)
-      return _children[key._elem]?._childrenInternal(key._prev) ??
+  Iterable<V> _childrenInternal([Path<K> key = const Path.empty()]) {
+    if (!key.isEmpty) {
+      return _children[key.first]?._childrenInternal(key.rest) ??
           Iterable.empty();
+    }
 
     final result = Set.of(_values);
     for (final child in _children.values) {
@@ -79,7 +91,9 @@ class PathMap<K, V> {
 
   Iterable<V> eachChildren(Iterable<Path<K>> paths) {
     final Set<V> result = Set.identity();
-    for (final path in paths) result.addAll(this.children(path));
+    for (final path in paths) {
+      result.addAll(children(path));
+    }
     return result;
   }
 
@@ -87,8 +101,6 @@ class PathMap<K, V> {
 
   Iterable<V> _getInternal(Path<K> key) {
     if (key.isEmpty) return _values;
-    if (_children.containsKey(key._elem))
-      return _children[key._elem]._getInternal(key);
-    return Iterable.empty();
+    return _children[key.first]?._getInternal(key.rest) ?? const [];
   }
 }
