@@ -1,44 +1,44 @@
 const overridable_operators = {
-  "<",
-  "+",
-  "|",
-  "[]",
-  ">",
-  "/",
-  "^",
-  "[]=",
-  "<=",
-  "~/",
-  "&",
-  "~",
-  ">=",
-  "*",
-  "<<",
-  "==",
-  "–",
-  "%",
-  ">>",
+  '<',
+  '+',
+  '|',
+  '[]',
+  '>',
+  '/',
+  '^',
+  '[]=',
+  '<=',
+  '~/',
+  '&',
+  '~',
+  '>=',
+  '*',
+  '<<',
+  '==',
+  '–',
+  '%',
+  '>>',
 };
 
 const binary_operators = {
-  "<",
-  "+",
-  "|",
-  ">",
-  "/",
-  "^",
-  "<=",
-  "~/",
-  "&",
-  ">=",
-  "*",
-  "<<",
-  "==",
-  "–",
-  "%",
-  ">>",
+  '<',
+  '+',
+  '|',
+  '>',
+  '/',
+  '^',
+  '<=',
+  '~/',
+  '&',
+  '>=',
+  '*',
+  '<<',
+  '==',
+  '–',
+  '%',
+  '>>',
 };
 
 const unary_prefix_operators = {
-  "~",
+  '~',
 };
